The Homepage of “Emergency Event Assistant” APP
We designed four functions (from left to right in the function bar): self-assessment (Occupational Heat Illness monitoring and warning system), popular science, prevention, and emergency (including text and video). When using a function, outdoor workers need to click the function key in the function bar first, and then select the high-temperature module to use the function. The homepage of the APP is shown in Figure 1.